functions separated into separate packages, documentation updated
