+++ docs/source/Utils.rst
@@ -1,0 +1,16 @@
+diPLSlib.utils
+==============
+
+
+.. automodule:: diPLSlib.utils.misc
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: diPLSlib.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
