+++ docs/source/conf.py
@@ -12,23 +12,37 @@ sys.path.insert(0, os.path.abspath('..'))
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = 'diPLSlib'
+project = 'diPLSlib Documentation'
 copyright = '2024, Ramin Nikzad-Langerodi'
 author = 'Ramin Nikzad-Langerodi'
-release = 'v2.0.0'
+release = '2.0.0'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
 extensions = [
     'sphinx.ext.autodoc',
-    'sphinx.ext.napoleon'
+    'sphinx.ext.autosummary',
+    'numpydoc'
+    # 'sphinx.ext.napoleon'
 ]
+
+autosummary_generate = True
 
 templates_path = ['_templates']
 exclude_patterns = []
 
-language = 'Python'
+language = 'en'
+
+autodoc_default_options = {
+    'member-order': 'bysource',
+    'undoc-members': True,
+    'show-inheritance': True,
+    'special-members': '__init__',
+    'exclude-members': '__weakref__',
+    'ignore-module-all': True,
+    'no-value': True
+}
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
