+++ docs/source/index.rst
@@ -1,22 +1,14 @@
-.. diPLSlib documentation master file, created by
-   sphinx-quickstart on Thu Oct 31 12:22:44 2024.
+.. diPLSlib Documentation documentation master file, created by
+   sphinx-quickstart on Sat Nov  2 18:56:24 2024.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to diPLSlib's documentation!
-====================================
+diPLSlib Documentation 
+======================
 
 
 .. toctree::
    :maxdepth: 2
    :caption: Contents:
-   :hidden:
 
    modules
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
