+++ docs/source/modules.rst
@@ -1,7 +1,8 @@
-diPLSlib
-========
+docs
+====
 
 .. toctree::
    :maxdepth: 4
 
    diPLSlib
+   Utils
